backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
ortools==9.9.3963
googlemaps==4.10.0
Pillow==10.2.0
django-filter==24.1 